refactor: remove endpoint that returns repo logs
